backend/requirements.txt
fastapi
uvicorn
langgraph
google-auth
google-auth-oauthlib
google-api-python-client
pydantic
python-dotenv
dateparser